Aligned code after core was renamed to engine.
