+++ tests/editable.js
@@ -7,8 +7,8 @@
 
 import Editor from '/ckeditor5/editor.js';
 import Editable from '/ckeditor5/editable.js';
-import EditingController from '/ckeditor5/core/treecontroller/editingcontroller.js';
-import ViewElement from '/ckeditor5/core/treeview/element.js';
+import EditingController from '/ckeditor5/engine/treecontroller/editingcontroller.js';
+import ViewElement from '/ckeditor5/engine/treeview/element.js';
 
 describe( 'Editable', () => {
 	const ELEMENT_NAME = 'h1';
